feat: add subscription prefix for admin and service to host
a host could have a scope in the database now; host could be linked to a specific service.
if this column is null the host would be global, therefor listed for users of all services who have that inbound, otherwise it's specific to a specific service and it's users.
the current method to list hosts for a user is subject to change